refactor: PaymentPage 컴포넌트에서 PaymentItem 분리
